% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flowproj.R
\name{plotflows}
\alias{plotflows}
\title{A function to plot flows in R.
This is a slight improvement over the RiverPlot package with slightly better display settings}
\usage{
plotflows(edges, spacing = 2, colorset = "Set1")
}
\arguments{
\item{edges:}{a matrix of format  N1 , N2 , Value}

\item{spacing:}{a number controlling the spacing of nodes defaults to 2}

\item{colorset:}{a parameter controlling which of the Rcolorbrewer colorsets to use}
}
\description{
A function to plot flows in R.
This is a slight improvement over the RiverPlot package with slightly better display settings
}
\examples{
edges = data.frame(N1 = paste0(rep(LETTERS[1:4], each = 4), rep(1:5, each = 16)),
N2 = paste0(rep(LETTERS[1:4], 4), rep(2:6, each = 16)),
Value = runif(80, min = 2, max = 5) * rep(c(1, 0.8, 0.6, 0.4, 0.3), each = 16),
stringsAsFactors = F)
plotflows(edges)
}
\keyword{RiverPlot}

